+++ gambler/source/Fold/Shortcut/Type.hs
@@ -1,7 +1,7 @@
 module Fold.Shortcut.Type
   (
     ShortcutFold (..),
-    Will (..), Vitality (..),
+    Will (..), Vitality (..), Vitality',
   )
   where
 
@@ -9,7 +9,7 @@ import Control.Applicative (Applicative, liftA2, pure, (<*>))
 import Data.Functor (Functor, fmap)
 import Data.Monoid (Monoid, mempty)
 import Data.Semigroup (Semigroup, (<>))
-import Strict (Will (..), Vitality (..))
+import Strict (Will (..), Vitality (..), Vitality')
 import Data.Void (absurd)
 
 import qualified Strict