+++ gambler/test/Spec/Nonempty.hs
@@ -4,10 +4,12 @@ import Fold.Nonempty
 
 import Test.Hspec
 
+import Control.Applicative (liftA2)
 import Data.Function ((&), flip)
 import Data.List.NonEmpty (NonEmpty ((:|)))
+import Data.Maybe (Maybe (..))
 import Positive (Positive)
-import Prelude (String, Integer)
+import Prelude (String, Integer, (+), odd)
 
 import qualified Data.List as List
 import qualified Fold.Pure as Pure
@@ -48,9 +50,26 @@ spec = describe "NonemptyFold" do
 
     describe "duplicate" do
         it "lets a fold run in two phases" do
-            let a, c :: NonEmpty Integer
+            let a :: NonEmpty Integer
                 b :: [Integer]
                 a = [1..3]
                 b = [4..6]
-                c = [1..6]
-            (sum & duplicate & flip run a & flip Pure.run b) `shouldBe` (List.sum c)
+            (sum & duplicate & flip run a & flip Pure.run b)
+                `shouldBe` (List.sum a + List.sum b)
+
+    describe "repeatedly" do
+        let a, b, c :: NonEmpty Integer
+            a = [2, 4]
+            b = [10, 12, 5, 7, 8]
+            c = [4, 5, 6]
+        it "can fold a list of lists" do
+            run (repeatedly run maximum) [a, b, c] `shouldBe` 12
+        it "works in Applicative combination with another fold" do
+            let f = liftA2 (,) (repeatedly run length) length
+            run f [a, b, c] `shouldBe` (10, 3)
+        it "works in Applicative combination with another 'repeatedly'" do
+            let f = liftA2 (,) (repeatedly run (find odd)) (repeatedly run maximum)
+            run f [a, b, c] `shouldBe` (Just 5, 12)
+        it "works on an Applicative combination" do
+            let f = liftA2 (,) (find odd) maximum
+            run (repeatedly run f) [a, b, c] `shouldBe` (Just 5, 12)